pyproject.toml
[tool.poetry]
name = "yolo-demo"
version = "0.1.0"
description = ""
authors = ["Iven Winkelmann <winkelmann.iven@gmail.com>"]
readme = "README.md"
packages = [{include = "yolo_demo"}]

[tool.poetry.dependencies]
python = "~3.9"
torch = "^2.0.1"
torchvision = "^0.15.2"
ultralytics = "^8.0.114"
paho-mqtt = "^1.6.1"
numpy = "~1.21.0"
lap = "^0.4.0"
shapely = "^2.0.1"
pafy = "^0.5.5"
youtube-dl = "^2021.12.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
